Just a lot changed, can render 20000 cubes fast now(30fps). Everything is duct taped together however. I will surely fix this.

+++ Data/Shader/ComputeShader.cpp
@@ -13,7 +13,7 @@ ComputeShader::ComputeShader(SHADER_FROM from, const char* compute)
 ComputeShader::~ComputeShader()
 {
 	glDeleteBuffers(this->shaderStorageBuffers.size(), this->shaderStorageBuffers.data());
-	for (size_t i = 0; i < this->shaderStorageBuffers.size(); i++) this->shaderStorageBuffers[i] = 0;
+	for (int i = 0; i < this->shaderStorageBuffers.size(); i++) this->shaderStorageBuffers[i] = 0;
 
 	glDeleteProgram(this->shaderProgramId);
